requirements.txt
# 智游助手v5.0 Python依赖包
# 基于DeepSeek API和高德MCP集成的旅游规划系统

# ================================
# 核心AI和语言模型 (Core AI & LLM)
# ================================

# OpenAI兼容客户端 (用于DeepSeek API)
openai>=1.0.0

# Token计数和文本处理
tiktoken>=0.5.0

# LangGraph和LangChain生态
langgraph>=0.0.40
langchain-core>=0.1.0
langchain>=0.1.0

# ================================
# 异步和网络请求 (Async & HTTP)
# ================================

# 异步HTTP客户端
aiohttp>=3.8.0

# 异步编程支持
asyncio-mqtt>=0.11.0  # 可选，用于消息队列
nest-asyncio>=1.5.0   # Jupyter异步支持

# 同步HTTP请求
requests>=2.28.0

# 重试机制
tenacity>=8.0.0

# ================================
# 数据处理和验证 (Data & Validation)
# ================================

# 数据模型和验证
pydantic>=2.0.0

# JSON处理增强
orjson>=3.8.0  # 可选，更快的JSON处理

# 数据分析
pandas>=1.5.0  # 可选，用于数据分析
numpy>=1.24.0  # 可选，数值计算

# ================================
# 模板和文档生成 (Templates & Docs)
# ================================

# 模板引擎
jinja2>=3.1.0

# Markdown处理
markdown>=3.4.0

# HTML处理
beautifulsoup4>=4.11.0  # 可选，HTML解析

# ================================
# 配置和环境管理 (Config & Environment)
# ================================

# 环境变量管理
python-dotenv>=1.0.0

# 配置文件处理
pyyaml>=6.0.0
toml>=0.10.0

# ================================
# 日志和监控 (Logging & Monitoring)
# ================================

# 结构化日志
structlog>=22.0.0  # 可选，结构化日志

# 性能监控
psutil>=5.9.0  # 可选，系统监控

# ================================
# 测试和开发工具 (Testing & Development)
# ================================

# 测试框架
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# 代码质量
black>=23.0.0  # 代码格式化
flake8>=6.0.0  # 代码检查
mypy>=1.0.0    # 类型检查

# ================================
# Jupyter和交互式开发 (Jupyter & Interactive)
# ================================

# Jupyter核心
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=6.5.0

# Jupyter扩展
ipywidgets>=8.0.0  # 交互式组件
matplotlib>=3.6.0  # 可选，图表绘制
plotly>=5.0.0      # 可选，交互式图表

# ================================
# 缓存和存储 (Cache & Storage)
# ================================

# 内存缓存
cachetools>=5.0.0

# Redis客户端 (可选)
redis>=4.5.0

# 数据库连接 (可选)
sqlalchemy>=2.0.0  # 可选，数据库ORM

# ================================
# 安全和加密 (Security & Encryption)
# ================================

# 加密和哈希
cryptography>=40.0.0
hashlib-compat>=1.0.0  # 可选，哈希兼容

# JWT处理 (可选)
pyjwt>=2.6.0

# ================================
# 时间和日期处理 (Time & Date)
# ================================

# 时区处理
pytz>=2023.0
python-dateutil>=2.8.0

# ================================
# 地理和地图相关 (Geography & Maps)
# ================================

# 地理计算 (可选)
geopy>=2.3.0
shapely>=2.0.0  # 可选，地理形状处理

# ================================
# 图像和媒体处理 (Image & Media)
# ================================

# 图像处理 (可选)
pillow>=9.5.0

# ================================
# 部署和生产环境 (Deployment & Production)
# ================================

# WSGI服务器 (可选)
gunicorn>=20.1.0

# 进程管理 (可选)
supervisor>=4.2.0

# ================================
# 开发依赖 (Development Dependencies)
# ================================

# 开发工具
pre-commit>=3.0.0  # Git钩子
commitizen>=3.0.0  # 提交规范

# 文档生成
sphinx>=6.0.0     # 可选，文档生成
mkdocs>=1.4.0     # 可选，Markdown文档

# ================================
# 特定版本锁定 (Version Pinning)
# ================================

# 确保兼容性的版本锁定
typing-extensions>=4.5.0
setuptools>=65.0.0
wheel>=0.40.0

# ================================
# 平台特定依赖 (Platform Specific)
# ================================

# Windows特定 (可选)
# pywin32>=306; sys_platform == "win32"

# macOS特定 (可选)
# pyobjc>=9.0; sys_platform == "darwin"

# Linux特定 (可选)
# python-systemd>=234; sys_platform == "linux"

# ================================
# 可选增强功能 (Optional Enhancements)
# ================================

# 性能优化
uvloop>=0.17.0; sys_platform != "win32"  # 更快的事件循环

# 内存优化
pympler>=0.9.0  # 可选，内存分析

# 并发处理
concurrent-futures>=3.1.0  # Python 3.8+兼容

# ================================
# 注释说明
# ================================

# 核心依赖 (必需):
# - openai: DeepSeek API客户端
# - tiktoken: Token计数
# - langgraph: 状态图引擎
# - aiohttp: 异步HTTP
# - pydantic: 数据验证
# - jinja2: 模板引擎
# - python-dotenv: 环境变量
# - nest-asyncio: Jupyter异步支持

# 可选依赖 (推荐):
# - pandas: 数据分析
# - redis: 缓存
# - pytest: 测试
# - jupyter: 交互式开发

# 开发依赖 (开发环境):
# - black: 代码格式化
# - mypy: 类型检查
# - pre-commit: Git钩子
